requirements.txt
transformers
diffusers
peft
opencv-python
protobuf
sentencepiece
gradio
jupyter
torchao
einops
matplotlib